swal migrated to single source for both front and backend
