+++ resources/assets/backend/backend-js.js
@@ -3,13 +3,12 @@ import "./backend-css.css";
 
 import.meta.glob(["../../images/**", "./images/**", "./fonts/**"]);
 
-/* Import sweetalert2 */
-import Swal from "sweetalert2";
-window.Swal = Swal;
-
 // import "./js/pace.min";
 /* Bootstrap */
 import "./js/bootstrap.bundle.min";
+
+/* Swal */
+import "../common/js/swal.js";
 
 /* Import jQuery */
 import $ from "jquery";
@@ -27,18 +26,3 @@ window.metisMenu = metisMenu;
 
 /* App */
 import "./js/app_backend";
-
-const Toast = Swal.mixin({
-    toast: true,
-    position: "bottom-left",
-    showConfirmButton: false,
-    timer: 3000,
-    timerProgressBar: true,
-    showCloseButton: true,
-    didOpen: (toast) => {
-        toast.addEventListener("mouseenter", Swal.stopTimer);
-        toast.addEventListener("mouseleave", Swal.resumeTimer);
-    },
-});
-
-window.Toast = Toast;
