+++ resources/assets/web/web-js.js
@@ -6,8 +6,12 @@ import.meta.glob(["./images/**", "../../images/**", "../fonts/**"]);
 import jQuery from "jquery";
 window.jQuery = window.$ = jQuery;
 
+/** Wow */
 import WOW from "wow.js";
 new WOW().init();
+
+/* Swal */
+import "../common/js/swal.js";
 
 import "../../../node_modules/boxicons/dist/boxicons";
 import "../../../node_modules/bootstrap/dist/js/bootstrap.bundle.min";
